app/requirements.txt
django==3.1.3
djangorestframework==3.12.2
django-webpack4-loader==0.0.5
gunicorn==20.0.4
whitenoise==5.2.0
django-filter==2.4.0